Fix CANcoder duplicate handles, NT legacy API, and int/long type mismatch
Issue 3: CalibrateCANcodersCommand now reads from existing SwerveModule
CANcoder handles via SwerveSubsystem.getCalibrationSamples() instead of
creating duplicate CANcoder objects on the same CAN IDs. This avoids
potential stale cache issues with Phoenix 6.

Issue 4: DashboardSnapshot matchNumber, canReceiveErrorCount, and
canTransmitErrorCount changed from int to long to match the DashboardData
record on the dashboard side and the NT IntegerPublisher contract.

Issue 5: CalibrateCANcodersCommand now uses typed DoublePublisher via
try-with-resources instead of legacy getEntry().setDouble() calls,
consistent with the rest of the codebase's NT usage.

+++ src/main/java/frc/robot/dashboard/DashboardSnapshot.java
@@ -45,7 +45,7 @@ public record DashboardSnapshot(
         String[] autoOptions,
         boolean autoCommandRunning,
         // Match info
-        int matchNumber,
+        long matchNumber,
         String eventName,
         // Camera / vision connection
         boolean cameraConnected,
@@ -54,8 +54,8 @@ public record DashboardSnapshot(
         double visionDistanceM,
         // CAN bus health
         double canBusUtilization,
-        int canReceiveErrorCount,
-        int canTransmitErrorCount,
+        long canReceiveErrorCount,
+        long canTransmitErrorCount,
         // Swerve module angles (FL, FR, BL, BR)
         double swerveFLAngleDeg,
         double swerveFRAngleDeg,
